+++ forui/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.21.0 (Next)
+
+### `FItem`
+* Add `FInheritedItemCallbacks`.
+* Add `FItemMixin.submenu(...)` shorthand for `FSubmenuItem`.
+
+
+### `FPopoverMenu`
+* Add `FSubmenuItem`.
+* Add `FSubmenuTile`.
+* Add `FPopoverMenu.hover`.
+* Add `FPopoverMenuStyle.motion`.
+* Add `FPopoverMenuMotion`.
+* Add `FPopoverMenuStyle.minWidth`.
+
+
+### `FTile`
+* Add `FTileMixin.submenu(...)` shorthand for `FSubmenuTile`.
+
+
+
+
 ## 0.20.0
 
 This update builds upon the styling overhaul by introducing desktop and touch variants of each theme and making widget